+++ imports/ui/components/advisories/advisories-content.html
@@ -1,0 +1,18 @@
+<template name="AdvisoriesContent">
+  
+  <div class="container advisory-header">
+    <h3>Advisories</h3>
+  </div>
+
+  <div>
+    {{#navybitsPagination data=advisories}}
+      <div class="container col">
+        <h4>{{title}}</h4>
+        <h6>{{{subtitle}}}</h6>
+        <div class="advisory-body">{{{body}}}</div>
+      </div>
+      <hr style="width: 70%;">
+    {{/navybitsPagination}}
+  </div>
+
+</template>
