+++ imports/ui/components/advisories/advisories-subheader.html
@@ -1,0 +1,20 @@
+<template name="AdvisoriesSubheader">
+
+  <nav class="navbar navbar-default nav-container">
+    <div id="sub-navbar" class="container-fluid">
+      <div class="navbar navbar-static-top" role="navigation">
+          <h4 id="advisory-monitoring" class="nav navbar-nav navbar-left">MONITORING</h4>
+          <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+            <ul class="nav navbar-nav navbar-right">
+              <li><a href="/weather-monitoring">Weather</a></li>
+              <li><a href="/heat-map-rainfall-outlook">Rainfall Outlook</a></li>
+              <li><a href="https://sarai-satellite-analysis.staging.lkexi.xyz/">Comparative Satellite Images</a></li>
+              <li><a href="/suitability-gallery">Suitability</a></li>
+              <li><a href="/advisories">Advisory</a></li>
+            </ul>
+          </div>
+      </div>
+    </div>
+  </nav>
+
+</template>
