+++ imports/ui/components/advisories/advisories-subheader.js
@@ -1,0 +1,2 @@
+import { Meteor } from 'meteor/meteor';
+import './advisories-subheader.html';
