+++ imports/ui/components/home/rainfall-outlook.js~77921cce6449b18df9b80f6df1f5b4945fadee22
@@ -1,0 +1,2 @@
+import { Meteor } from 'meteor/meteor';
+import './rainfall-outlook.html';
