+++ imports/ui/components/home/rainfall-outlook.js~Added Partner institutions in About page
@@ -1,0 +1,2 @@
+import { Meteor } from 'meteor/meteor';
+import './rainfall-outlook.html';
